Move inquiry stream to ee
